requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3.1
cssselect==1.0.3
d==0.2.2
Django==1.11.15
django-cors-headers==2.4.0
django-crispy-forms==1.7.0
django-webpack-loader==0.2.4
djangorestframework==3.8.2
enum34==1.1.6
Fabric==1.10.2
gyp==0.1
idna==2.7
ipaddress==1.0.22
keyring==10.6.0
keyrings.alt==3.0
lxml==4.2.4
Markdown==2.6.11
paramiko==2.4.1
pbr==4.2.0
pyasn1==0.4.4
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
PyNaCl==1.2.1
pyquery==1.4.0
pytz==2018.5
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
sqlparse==0.2.4
stevedore==1.29.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
